moved swipeable carousel to seperate dir for future reuse
