+++ src/component/SwipeableCarousel/SwipeableCarousel.css
