+++ src/component/SwipeableCarousel/SwipeableCarousel.jsx
